Move DispatchOnMainQueueDecorator to production target
